Backend/services/ai_integration.py
import os
import asyncio
import json
from typing import Dict, Any, List
import requests
from google.oauth2 import service_account
from googleapiclient.discovery import build
from dotenv import load_dotenv

# Load environment variables
load_dotenv()

# This would be replaced with actual Gemini API credentials in production
GEMINI_API_KEY = os.getenv("GEMINI_API_KEY", "your-api-key")

async def generate_cut_profile(video_path: str, transcript_path: str) -> Dict[str, Any]:
    """
    Use Gemini API to generate a smart cut profile based on video and transcript.
    
    Args:
        video_path: Path to the video file
        transcript_path: Path to the transcript file
        
    Returns:
        Dictionary containing cut points and editing instructions
    """
    # Read the transcript file
    with open(transcript_path, 'r') as f:
        transcript_text = f.read()
    
    # For demo purposes, we'll use a simpler API call
    # In a real implementation, you would:
    # 1. Extract key frames from the video
    # 2. Send both transcript and key frames to Gemini
    # 3. Process the response to create cut points
    
    # Simple Gemini API request via REST API
    prompt = f"""
    You are a professional video editor. Based on the following transcript, 
    create a cut profile for a video. Identify natural breaks, redundant content, 
    and awkward pauses that should be removed. Create a JSON output with segments 
    to keep in the final video.
    
    Transcript:
    {transcript_text}
    
    Respond with a JSON containing:
    1. An array of segments, each with start_time and end_time in seconds
    2. Each segment should also have a reason field explaining why this part should be kept
    3. Provide an estimated final duration
    """
    
    # For demonstration, we'll simulate a response
    # In production, you would use the actual Gemini API
    if os.getenv("USE_MOCK_API", "true").lower() == "true":
        # Mock response for demonstration
        # In a real implementation, this would come from the Gemini API
        mock_response = {
            "segments": [
                {
                    "start_time": 0.0,
                    "end_time": 45.3,
                    "reason": "Strong opening hook and introduction of main topic"
                },
                {
                    "start_time": 62.1,
                    "end_time": 125.7,
                    "reason": "Key explanation of core concept with good energy"
                },
                {
                    "start_time": 140.5,
                    "end_time": 210.2,
                    "reason": "Important demonstration with clear visual elements"
                },
                {
                    "start_time": 250.8,
                    "end_time": 320.4,
                    "reason": "Compelling conclusion and call to action"
                }
            ],
            "estimated_duration": 248.2,
            "notes": "Cut removes redundant explanations and several awkward pauses. Maintains narrative flow while improving pacing."
        }
        
        await asyncio.sleep(2)  # Simulate API latency
        return mock_response
    else:
        # Real Gemini API call
        # This is a placeholder - actual implementation would use the official Gemini API
        api_url = "https://generativelanguage.googleapis.com/v1beta/models/gemini-pro:generateContent"
        
        headers = {
            "Content-Type": "application/json",
            "x-goog-api-key": GEMINI_API_KEY
        }
        
        payload = {
            "contents": [
                {
                    "parts": [
                        {"text": prompt}
                    ]
                }
            ]
        }
        
        response = requests.post(api_url, headers=headers, json=payload)
        
        if response.status_code != 200:
            raise Exception(f"Gemini API error: {response.status_code} - {response.text}")
        
        # Parse the response to extract the JSON cut profile
        # The actual parsing would depend on the exact response format from Gemini
        result = response.json()
        text_response = result["candidates"][0]["content"]["parts"][0]["text"]
        
        # Extract JSON from the text response
        import re
        json_match = re.search(r'```json\n(.*?)\n```', text_response, re.DOTALL)
        if json_match:
            cut_profile_json = json_match.group(1)
        else:
            # Try to find any JSON-like structure
            json_match = re.search(r'(\{.*\})', text_response, re.DOTALL)
            if json_match:
                cut_profile_json = json_match.group(1)
            else:
                raise Exception("Could not extract JSON from Gemini response")
        
        cut_profile = json.loads(cut_profile_json)
        return cut_profile
