Backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.23.2
python-multipart>=0.0.6
moviepy>=1.0.3
google-api-python-client>=2.100.0
sqlalchemy>=2.0.22
pydantic>=2.4.2
python-dotenv>=1.0.0
requests>=2.31.0
numpy>=1.26.0
pytest>=7.4.2
httpx>=0.25.0
webvtt-py>=0.4.6
